feat: alerts check_alerts loop, table layout, Polygon volume f64

+++ src/models/historical.rs
@@ -2,10 +2,15 @@ use serde::Deserialize;
 
 #[derive(Deserialize, Debug)]
 pub struct HistoricalResponse {
+    #[serde(default)]
     pub ticker: String,
+    #[serde(default)]
     pub results: Vec<HistoricalData>,
+    #[serde(default)]
     pub status: String,
+    #[serde(default)]
     pub request_id: String,
+    #[serde(default)]
     pub count: u32,
 }
 
@@ -15,7 +20,7 @@ pub struct HistoricalData {
     pub h: f64, // High
     pub l: f64, // Low
     pub c: f64, // Close
-    pub v: u64, // Volume
+    pub v: f64, // Volume (Polygon may return fractional values)
     pub t: u64, // Timestamp
     pub vw: f64, // Volume weighted average price
     pub n: Option<u64>, // Number of transactions
